requirements.txt
autopep8==1.5.7
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
neo4j==4.2.1
numpy==1.20.2
pandas==1.2.4
psycopg2-binary==2.8.6
pycodestyle==2.7.0
PyJWT==2.1.0
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
SQLAlchemy==1.4.11
toml==0.10.2
Werkzeug==1.0.1
